Faire en sorte que seul l'utilisateur ayant le mail admin@admin.com puisse ajouter, modifier ou supprimer des produits. Une solution simple et générique devra être utilisée. Il n'est pas nécessaire de mettre en place une gestion des accès basée sur les rôles.

+++ back/src/main/java/com/ecommerce/alten/services/ProductService.java
@@ -3,6 +3,7 @@ package com.ecommerce.alten.services;
 import com.ecommerce.alten.models.Product;
 import com.ecommerce.alten.repositories.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,10 +22,16 @@ public class ProductService {
     }
 
     public Product createProduct(Product product) {
+        if (isAdmin()) {
+            throw new SecurityException("Unauthorized: Only admin can delete products.");
+        }
         return productRepository.save(product);
     }
 
     public Product updateProduct(Long id, Product product) {
+        if (isAdmin()) {
+            throw new SecurityException("Unauthorized: Only admin can delete products.");
+        }
         Product existingProduct = productRepository.findById(id).orElse(null);
         if (existingProduct != null) {
             existingProduct.setName(product.getName());
@@ -36,6 +43,15 @@ public class ProductService {
     }
 
     public void deleteProduct(Long id) {
+        if (isAdmin()) {
+            throw new SecurityException("Unauthorized: Only admin can delete products.");
+        }
         productRepository.deleteById(id);
+    }
+
+    private boolean isAdmin() {
+        String currentUserEmail = SecurityContextHolder.getContext().getAuthentication().getName();
+        System.out.println("le mail "+currentUserEmail);
+        return !"admin@admin.com".equals(currentUserEmail);
     }
 }
